Q&A_Chatbot_with_FireworksAI/requirements.txt
langchain-fireworks
langchain
python-dotenv
langchain-core
streamlit
fireworks-ai